requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
attrs==23.1.0
    # via
    #   outcome
    #   trio
certifi==2023.11.17
    # via
    #   requests
    #   selenium
charset-normalizer==3.3.2
    # via requests
exceptiongroup==1.2.0
    # via
    #   trio
    #   trio-websocket
h11==0.14.0
    # via wsproto
idna==3.6
    # via
    #   requests
    #   trio
outcome==1.3.0.post0
    # via trio
packaging==23.2
    # via webdriver-manager
pysocks==1.7.1
    # via urllib3
python-dotenv==1.0.0
    # via webdriver-manager
requests==2.31.0
    # via webdriver-manager
selenium==4.16.0
    # via -r requirements.in
sniffio==1.3.0
    # via trio
sortedcontainers==2.4.0
    # via trio
trio==0.23.2
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
urllib3[socks]==2.1.0
    # via
    #   requests
    #   selenium
    #   urllib3
webdriver-manager==4.0.1
    # via -r requirements.in
wsproto==1.2.0
    # via trio-websocket
